# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BuildRMC Fortune Dashboard API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales-orders:
    get:
      operationId: getSalesOrders
      tags: [sales]
      summary: List all sales orders
      responses:
        "200":
          description: List of sales orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesOrder"
    post:
      operationId: createSalesOrder
      tags: [sales]
      summary: Create a new sales order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesOrderBody"
      responses:
        "201":
          description: Sales order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrder"
  /sales-orders/{id}:
    get:
      operationId: getSalesOrder
      tags: [sales]
      summary: Get a sales order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sales order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrder"
    put:
      operationId: updateSalesOrder
      tags: [sales]
      summary: Update a sales order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalesOrderBody"
      responses:
        "200":
          description: Sales order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesOrder"
    delete:
      operationId: deleteSalesOrder
      tags: [sales]
      summary: Delete a sales order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /qc-tests:
    get:
      operationId: getQcTests
      tags: [qc]
      summary: List all QC tests
      responses:
        "200":
          description: List of QC tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QcTest"
    post:
      operationId: createQcTest
      tags: [qc]
      summary: Create a new QC test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQcTestBody"
      responses:
        "201":
          description: QC test created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcTest"
  /qc-tests/{id}:
    get:
      operationId: getQcTest
      tags: [qc]
      summary: Get a QC test by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QC test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcTest"
    put:
      operationId: updateQcTest
      tags: [qc]
      summary: Update a QC test
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQcTestBody"
      responses:
        "200":
          description: QC test updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcTest"
    delete:
      operationId: deleteQcTest
      tags: [qc]
      summary: Delete a QC test
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /products:
    get:
      operationId: getProducts
      tags: [store]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [store]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [store]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [store]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [store]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /inventory-entries:
    get:
      operationId: getInventoryEntries
      tags: [store]
      summary: List inventory entries
      responses:
        "200":
          description: List of inventory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryEntry"
    post:
      operationId: createInventoryEntry
      tags: [store]
      summary: Create an inventory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryEntryBody"
      responses:
        "201":
          description: Inventory entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryEntry"
  /vehicles:
    get:
      operationId: getVehicles
      tags: [transport]
      summary: List all vehicles
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [transport]
      summary: Create a new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "201":
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [transport]
      summary: Get a vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    put:
      operationId: updateVehicle
      tags: [transport]
      summary: Update a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "200":
          description: Vehicle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [transport]
      summary: Delete a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips:
    get:
      operationId: getTrips
      tags: [transport]
      summary: List all trips
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [transport]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripBody"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
...
components:
  schemas:
...
    SalesOrder:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        orderDate:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        totalAmount:
          type: number
        status:
          type: string
        deliveryDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - orderNumber
        - orderDate
        - customerId
        - customerName
        - totalAmount
        - status
        - createdAt
    CreateSalesOrderBody:
      type: object
      properties:
        orderNumber:
          type: string
        orderDate:
          type: string
        customerId:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
        deliveryDate:
          type: ["string", "null"]
      required:
        - orderNumber
        - orderDate
        - customerId
        - totalAmount
        - status
    QcTest:
      type: object
      properties:
        id:
          type: integer
        testDate:
          type: string
        sampleCode:
          type: string
        slump:
          type: number
        strength:
          type: number
        grade:
          type: string
        status:
          type: string
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - testDate
        - sampleCode
        - slump
        - strength
        - grade
        - status
        - createdAt
    CreateQcTestBody:
      type: object
      properties:
        testDate:
          type: string
        sampleCode:
          type: string
        slump:
          type: number
        strength:
          type: number
        grade:
          type: string
        status:
          type: string
        remarks:
          type: ["string", "null"]
      required:
        - testDate
        - sampleCode
        - slump
        - strength
        - grade
        - status
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        unitPrice:
          type: number
        stockQty:
          type: number
        minStockLevel:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - unit
        - unitPrice
        - stockQty
        - minStockLevel
        - createdAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        unitPrice:
          type: number
        stockQty:
          type: number
        minStockLevel:
          type: number
      required:
        - name
        - category
        - unit
        - unitPrice
        - stockQty
        - minStockLevel
    InventoryEntry:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        entryDate:
          type: string
        quantity:
          type: number
        type:
          type: string
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - entryDate
        - quantity
        - type
        - createdAt
    CreateInventoryEntryBody:
      type: object
      properties:
        productId:
          type: integer
        entryDate:
          type: string
        quantity:
          type: number
        type:
          type: string
        remarks:
          type: ["string", "null"]
      required:
        - productId
        - entryDate
        - quantity
        - type
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        registrationNo:
          type: string
        model:
          type: string
        capacity:
          type: number
        status:
          type: string
        driverName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - registrationNo
        - model
        - capacity
        - status
        - createdAt
    CreateVehicleBody:
      type: object
      properties:
        registrationNo:
          type: string
        model:
          type: string
        capacity:
          type: number
        status:
          type: string
        driverName:
          type: ["string", "null"]
      required:
        - registrationNo
        - model
        - capacity
        - status
    Trip:
      type: object
      properties:
        id:
          type: integer
        tripDate:
          type: string
        vehicleId:
          type: integer
        vehicleReg:
          type: string
        destination:
          type: string
        loadQty:
          type: number
        status:
          type: string
        driverName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - tripDate
        - vehicleId
        - vehicleReg
        - destination
        - loadQty
        - status
        - createdAt
    CreateTripBody:
      type: object
      properties:
        tripDate:
          type: string
        vehicleId:
          type: integer
        destination:
          type: string
        loadQty:
          type: number
        status:
          type: string
        driverName:
          type: ["string", "null"]
      required:
        - tripDate
        - vehicleId
        - destination
        - loadQty
        - status
...