# lib/api-zod/src/generated/types/createVehicleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildRMC Fortune Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVehicleBody {
  registrationNo: string;
  model: string;
  capacity: number;
  status: string;
  /** @nullable */
  driverName?: string | null;
}