# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildRMC Fortune Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  category: string;
  unit: string;
  unitPrice: number;
  stockQty: number;
  minStockLevel: number;
  createdAt: string;
}