# lib/api-zod/src/generated/types/salesOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildRMC Fortune Dashboard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SalesOrder {
  id: number;
  orderNumber: string;
  orderDate: string;
  customerId: number;
  customerName: string;
  totalAmount: number;
  status: string;
  /** @nullable */
  deliveryDate?: string | null;
  createdAt: string;
}